Adapt to SF4

+++ .travis/before_install.php
@@ -3,5 +3,16 @@
 
 include_once 'common.php';
 
-// Disable XDebug
-run('phpenv config-rm xdebug.ini');
+// Remove memory limit on PHP 5
+if (isPhp5()) {
+        run(sprintf('echo "memory_limit=-1" >> ~/.phpenv/versions/%s/etc/conf.d/travis.ini;', getPhpVersion()));
+    }
+
+// Disable XDebug for non-experimental PHP environments
+if (isNonExperimentalPhp()) {
+        run('phpenv config-rm xdebug.ini');
+    }
+
+if (shouldBuildDocs()) {
+      //  run('sudo apt-get -qq update');
+    }
